Fixed the sizing for seeing the definition of the card
I fixed the container size for the term and definition swapping
to have a higher clickable area. In addition, I have noticed
there is a bug with the database when I cannot create another
card with the same name because it's already in a different container.
I will fix that tomorrow. *sleep*